Add game mode selector to lobby menu
