+++ client/src/components/gameModeSelector.css
@@ -1,0 +1,33 @@
+
+.saved-game-modes {
+    max-height: 50vh;
+    overflow-y: auto;
+    display: flex;
+    gap: .25rem;
+    flex-direction: row;
+    flex-wrap: wrap;
+    margin-top: 0.25rem;
+}
+.saved-game-modes > div {
+    display: flex;
+    flex-basis: 40%;
+    flex-grow: 1;
+    border: .13rem solid var(--primary-border-color);
+    background-color: var(--secondary-color);
+    border-radius: 0.5rem;
+    align-items: center;
+    justify-content: space-between;
+}
+.saved-game-modes > div > span {
+    margin: 0 .25rem;
+}
+.saved-game-modes > div > span.game-mode-name {
+    flex-grow: 1;
+}
+.saved-game-modes > div > div {
+    display: flex;
+    flex-direction: row;
+}
+.saved-game-modes > div > div > button {
+    margin: .1rem;
+}
