+++ client/src/components/gameModesEditor.css
@@ -14,37 +14,8 @@
     gap: 0.5rem;
 }
 
-.game-modes-editor .saved-game-modes {
-    max-height: 50vh;
-    overflow-y: auto;
-    display: flex;
-    gap: .25rem;
-    flex-direction: row;
-    flex-wrap: wrap;
-    margin-bottom: 0.25rem;
-}
 .game-modes-editor input {
     min-width: 3rem;
-}
-.game-modes-editor .saved-game-modes > div {
-    display: flex;
-    flex-basis: 40%;
-    flex-grow: 1;
-    border: .13rem solid var(--primary-border-color);
-    background-color: var(--secondary-color);
-    border-radius: 0.5rem;
-    align-items: center;
-    justify-content: space-between;
-}
-.game-modes-editor .saved-game-modes > div > span {
-    margin: 0 .25rem;
-}
-.game-modes-editor .saved-game-modes > div > div {
-    display: flex;
-    flex-direction: row;
-}
-.game-modes-editor .saved-game-modes > div > div > button {
-    margin: .1rem;
 }
 
 .game-modes-editor > main {
