+++ client/src/menu/lobby/lobbyMenu.css
@@ -12,13 +12,14 @@
     display: flex;
     flex-direction: row;
     justify-content: space-between;
-    align-items: baseline;
+    align-items: center;
     flex-wrap: wrap;
     gap: 0.5rem;
 }
 
 .lm > header input {
     font-size: 1.5rem;
+    
 }
 
 @media only screen and (max-width: 600px) {
